Extracted the interface for app preferences
