+++ app/src/main/java/com/omkarmoghe/pokemap/app_preferences/PokemapSharedPreferences.java
@@ -1,4 +1,4 @@
-package com.omkarmoghe.pokemap.utils;
+package com.omkarmoghe.pokemap.app_preferences;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -6,10 +6,10 @@ import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 
 /**
- * ToDo: Explain the scope of this class
+ * Provide convenience methods to access shared preferences
  */
 
-public final class PokemapSharedPreferences {
+public final class PokemapSharedPreferences implements PokemapAppPreferences {
     private static final String USERNAME_KEY = "UsernameKey";
     private static final String PASSWORD_KEY = "PasswordKey";
 
@@ -19,27 +19,33 @@ public final class PokemapSharedPreferences {
         sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
     }
 
+    @Override
     public boolean isUsernameSet() {
         return sharedPreferences.contains(USERNAME_KEY);
     }
 
+    @Override
     public boolean isPasswordSet() {
         return sharedPreferences.contains(PASSWORD_KEY);
     }
 
 
+    @Override
     public String getUsername() {
         return sharedPreferences.getString(USERNAME_KEY, "");
     }
 
+    @Override
     public void setUsername(@NonNull String username) {
         sharedPreferences.edit().putString(USERNAME_KEY, username).apply();
     }
 
+    @Override
     public void setPassword(@NonNull String password) {
         sharedPreferences.edit().putString(PASSWORD_KEY, password).apply();
     }
 
+    @Override
     public String getPassword() {
         return sharedPreferences.getString(PASSWORD_KEY, "");
     }
