+++ app/src/test/java/com/omkarmoghe/pokemap/app_preferences/PokemapSharedPreferencesTest.java
@@ -1,14 +1,15 @@
-package com.omkarmoghe.pokemap.utils;
+package com.omkarmoghe.pokemap.app_preferences;
 
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
+import com.omkarmoghe.pokemap.app_preferences.PokemapSharedPreferences;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
